requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
asyncpg==0.30.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
crawl4ai>=0.7.8
fonttools==4.55.6
greenlet==3.1.1
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
h2==4.1.0
hiredis==3.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
kiwisolver==1.4.8
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
openai==1.59.3
packaging==24.2
passlib==1.7.4
pillow==11.1.0
playwright==1.49.1
portalocker==2.10.1
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.2
qdrant-client==1.12.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
scipy==1.15.1
sentence-transformers>=2.2.0
gliner>=0.2.0
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
tenacity==9.0.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
wheel==0.44.0
